Finally - GetMemoryMap verifiably working correctly.

+++ kernel.h
@@ -3,9 +3,9 @@
 
 #include "graphics.h"
 
-void drawNum(int num, int *line, int *column);
+void drawNum(uint64_t num, int *line, int *column);
 
-void kprintf(const char *output, int num);
+void kprintf(const char *output, uint64_t num);
 
 void kprint(const char *output);
